Create get_user_and_location.sh
My techs weren't going back in to the JSS and assigning computers to users, so I had to find a way to automate the process. Other Casper Admins have developed similar things. I have this set to run once a day.

Keep in mind that you have to have "Collect user and location information from LDAP" enabled in Inventory Collection in your JSS.